// InvoiceSystem/Frontend/invoice-app/src/app/models/invoice.ts
export interface Invoice {
  id: number;
  invoiceNumber: string;
  invoiceDate: Date;
  customer: {
    name: string;
    email: string;
    phone: string;
    address: string;
  };
  items: InvoiceItem[];
  subTotal: number;
  discount: number;
  total: number;
}

export interface InvoiceItem {
  id?: number; // Make id optional since it's auto-generated by backend
  productId: number;
  productName: string;
  description: string;
  quantity: number;
  price: number;
  amount: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  stock?: number;
}